Restructure notebooks into per-topic folders; 2RN2 steered ΔG example
Reorganise notebooks/ into self-documented subfolders (Enhanced_Diffusion_
Sampling_Protein, Enhanced_Diffusion_Sampling_GMM, Physicality_Steering, PPFT),
each with its own README, and slim notebooks/README.md to a brief index.

The protein steered free-energy example is consolidated for release as a
two-file 2RN2 example:
- steered_dg_2rn2.py: single-run, single-GPU generation of ~1000 steered and
  unsteered samples, then a ΔG convergence curve (with error bars) via batch
  subsampling.
- steered_dg_lib.py: shared library (2RN2 system, FKC steering / unsteered
  denoiser configs, FNC ΔG recipe, subsampling utilities).

FKC RMSD steering recovers the internal reference folding ΔG (≈ -5.3 kcal/mol)
at a sample budget where the unsteered baseline saturates and cannot resolve the
rare unfolded state.
