+++ notebooks/Enhanced_Diffusion_Sampling_GMM/gmm_umbrella_mbar.py
