+++ notebooks/PPFT/README.md
@@ -1,0 +1,7 @@
+# Property-prediction fine-tuning (`PPFT/`)
+
+## Files
+
+- **`ppft_example.ipynb`** — Notebook demonstrating fine-tuning a pretrained BioEmu model with the PPFT (property-prediction fine-tuning) loss to shift the folded-state population for a single protein.
+- **`rollout.yaml`** — Hydra config for the denoising rollout used by the PPFT example.
+- **`HHH_rd1_0335.pdb`** — Reference structure used by the PPFT notebook (downloaded from https://zenodo.org/records/7992926).
