+++ notebooks/Physicality_Steering/README.md
@@ -1,0 +1,5 @@
+# Physical (clash) steering (`Physicality_Steering/`)
+
+## Files
+
+- **`physical_steering_example.py`** — Minimal comparison of sampling with and without physicality steering for a single sequence, using `config/steering/physical_steering.yaml`. Good starting point for the steered vs. unsteered `bioemu.sample.main` pattern. Run with `python physical_steering_example.py`.
